+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Merge upstream changes from original repo into this fork
+- Revert changes made to the handler as they are not needed
+- Extend SessionStore to return all sessions
+
+### Fixed
+- Use the right ID when deleting session on logout
 
 ## [v2.3.1] - 2025-10-09
 ### Fixed